Chart.js Libraries
Tried installing the chart.js libraries.
Libraries installed, but chart still not displaying on page.
Delete the fusion chart library as no longer needed.
Created a new view just for the chart.
Changed the jscharts script to include a new function.

Chart still needs work as not displaying on page as of yet.