Made the chart.js library output charts
Ive managed to output a chart based on some components
of the database, set in the pages controller.

The chart shows various brands and the models
assoisiated with them, showing which has the most.

Also tidied up some of the layout of the model show view
as well as the watch brand show view.