Renamed the charts view to dashboard
Renamed all instances of thec chart view to dashboard
and update the view with two more charts, a line
and a pie chart. Indented some of the code and cleaned
up some of the css for the charts backgrounds.

Added a dashboard link in the navbar to show the dashboard.